+++ CHANGELOG.md
@@ -8,12 +8,15 @@ PRs and issues that were included in a particular release.
 
 3.9 (in development)
 --------------------
+* Add MobilizedBody::CantileverFreeBeam, a mobilizer for modeling spinal joints and
+  other flexible structures. (PR #844)
+* Fixed implementations of setUToFitAngularVelocityImpl for the gimbal and bushing
+  mobilizers. (PR #844)
 * Add LLT (Cholesky) factorization
 * Add quaternion multiplication (Hamilton product) 
 * Required C++ level was increased from C++11 to C++17
 * Fixed an out-of-bounds warning/error emitted by overloads of SimTK::Mat accessing
   out-of-bounds rows
-
 
 3.8 (May 2025)
 --------------------